fixed active source switching for single number
previous code would result in 3X:00, i.e. physical address 3.x.0.0
This is not what would be expected in most setups.
First number is always the HDMI input of TV (here 3). The following pars
are used with switches only. So the old code only works for people who
have a switch on input three at their TVs.
But most people would expect the TV to switch HDMI Input, not the
switch...