fixed raw command state
previously did crash with "no method cec.sendCommand", which is true.
cec.send should do the trick quite fine.
One possibility would be to add "tx " for strings that are only CEC
syntax (i.e. DD:DD:DD) so people could work with cecomatic (for example)
and wouldn't need to know the syntax of cec-client.
I'll remember that as todo for later.