ignore active_source states that are not string
for some reason I get a 0 (number) here when I switch of my multimedia
setup which leads to exceptions in the log file. I do not really know
where the number value comes from, but I think the type check is not
bad, because we need a string to call "split" on it anyways.

If there is a scenario where we would want to process other types of
variables, we would still need to implement that anyway.